Restructure config test infrastructure
* move common helpers into own file
* isolate set of repo interface tests
* use interface tests in postgres tests
